+++ cmake/BipedalLocomotionControllersFindDependencies.cmake
@@ -123,9 +123,6 @@ checkandset_dependency(OsqpEigen)
 find_package(iDynTree 0.11.105 QUIET)
 checkandset_dependency(iDynTree)
 
-find_package(Catch2 QUIET)
-checkandset_dependency(Catch2)
-
 find_package(YARP QUIET)
 checkandset_dependency(YARP)
 
@@ -136,6 +133,10 @@ bipedal_locomotion_controllers_dependent_option(BIPEDAL_LOCOMOTION_CONTROLLERS_C
 bipedal_locomotion_controllers_dependent_option(BIPEDAL_LOCOMOTION_CONTROLLERS_COMPILE_YarpImplementation
   "Compile All the YARP implementations?" ON
   "BIPEDAL_LOCOMOTION_CONTROLLERS_COMPILE_YarpUtilities" OFF)
+
+bipedal_locomotion_controllers_dependent_option(BIPEDAL_LOCOMOTION_CONTROLLERS_COMPILE_Estimators
+  "Compile Estimators library?" ON
+  "BIPEDAL_LOCOMOTION_CONTROLLERS_HAS_iDynTree" OFF)
 
 bipedal_locomotion_controllers_dependent_option(BIPEDAL_LOCOMOTION_CONTROLLERS_COMPILE_ContactModels
   "Compile ContactModels library?" ON
@@ -153,6 +154,4 @@ bipedal_locomotion_controllers_dependent_option(BIPEDAL_LOCOMOTION_CONTROLLERS_C
   "Compile WholebodyControllers library?" ON
   "BIPEDAL_LOCOMOTION_CONTROLLERS_HAS_iDynTree;BIPEDAL_LOCOMOTION_CONTROLLERS_COMPILE_ContactModels;BIPEDAL_LOCOMOTION_CONTROLLERS_COMPILE_Simulator" OFF)
 
-bipedal_locomotion_controllers_dependent_option(BIPEDAL_LOCOMOTION_CONTROLLERS_COMPILE_tests
-  "Compile tests?" ON
-  "BIPEDAL_LOCOMOTION_CONTROLLERS_HAS_Catch2;BUILD_TESTING" OFF)
+