+++ CHANGELOG.md
@@ -13,5 +13,7 @@ All notable changes to this project are documented in this file.
 - Implement the  first version of the `YarpUtilities` library
 - Implement `ParametersHandler` library (https://github.com/dic-iit/bipedal-locomotion-controllers/pull/13)
 - Implement the first version of the `OptimalControlUtilities` library
+- Implement `GenericContainer::Vector` (https://github.com/dic-iit/bipedal-locomotion-controllers/pull/29)
+- Implement `Estimators` library (https://github.com/dic-iit/bipedal-locomotion-controllers/pull/23)
 
 [Unreleased]: https://github.com/dic-iit/bipedal-locomotion-controllers/